Add status to audio

+++ app/controllers/RaspController.php
@@ -15,6 +15,7 @@ class RaspController extends BaseController
 	private $playerOutput;
 	private $command;
 	private $switchStatus = array();
+	private $setMixer= 'amixer cset numid=3 1';
 	
 	
 //sudo chmod 4755 /usr/sbin/i2cdetect /usr/sbin/i2cset /usr/sbin/i2cget /usr/sbin/i2cdump
@@ -30,14 +31,12 @@ class RaspController extends BaseController
 
 	public function prepareData($data)
 	{
-		$process = substr($data, 0, strpos($data, "&"));
-		$this->switchStatus['status'] = strstr($data, 'status');		
+		$process = substr($data, 0, strpos($data, "&"));	
 		$tempAudioFile = strstr($data,'&');
 		$tempAudioFile = str_replace('=','',strstr($tempAudioFile, '='));
 		$this->audioFile = substr($tempAudioFile, 0, strpos($data, "&"));
 		$this->audioFile = substr($this->audioFile, 0, strpos($this->audioFile, "&"));
-		$this->switchStatus['name'] = '';
-		
+		$this->switchStatus[$this->audioFile] = strstr($data, 'status');
 		return $this->getCommand($process);
 	}
 
@@ -52,11 +51,8 @@ class RaspController extends BaseController
 		$hexaddress = '0x'.(dechex(bindec($toHex)));
 		$this->command = '/usr/sbin/i2cset -y 1 '.$comProcess.' '.$hexaddress;
 		$this->executeProcess();
-		if ($this->switchStatus['status'] == 'status0'){
-			return $this->playAudio($this->audioFile);
-		} else if ($this->switchStatus['name'] == $this->audioFile){
-			return $this->killProcess('pidof mpg123 | xargs kill -9');
-		}
+		
+		return $this->playAudio($this->audioFile);	
 	}
 
 	public function executeProcess()
@@ -72,12 +68,17 @@ class RaspController extends BaseController
 
 	public function playAudio($command) 
 	{	
-		$this->killProcess('pidof mpg123 | xargs kill -9');	
-		$this->switchStatus['name'] = $command;
+	    if ($this->switchStatus[$command] == 'status1'){
+            $this->killProcess('pidof mpg123 | xargs kill -9');
+        }
+
+
 		$this->audioprocess = new Process('mpg123 media/'.$command);
+//        return $this->switchStatus[$command]; //development
 		try {
 			$this->audioprocess->mustRun();
-			return $this->playerOutput = $this->audioprocess->getOutput();
+//			return $this->audioprocess->isRunning(); // development
+			 return $this->playerOutput = $this->audioprocess->getOutput();
 		} catch(ProcessFailedException $e) {
 			return $this->output = $e->getMessage();
 		}
